+++ lib/avx512_t2/mb_mgr_avx512_t2.c
@@ -263,8 +263,8 @@
 #define FLUSH_JOB_ZUC_NIA6  flush_job_zuc_nia6_gfni_avx512
 
 /* ZUC NCA6 */
-#define SUBMIT_JOB_ZUC_NCA6 submit_job_zuc_nca6_gfni_sse
-#define FLUSH_JOB_ZUC_NCA6  flush_job_zuc_nca6_gfni_sse
+#define SUBMIT_JOB_ZUC_NCA6 submit_job_zuc_nca6_gfni_avx512
+#define FLUSH_JOB_ZUC_NCA6  flush_job_zuc_nca6_gfni_avx512
 
 /* SNOW5G-NEA4 */
 #define SUBMIT_JOB_SNOW5G_NEA4_X2 submit_job_snow5g_nea4_vaes_avx512
@@ -429,8 +429,8 @@ reset_ooo_mgrs(IMB_MGR *state)
         ooo_mgr_zuc_reset(state->zuc_eia3_ooo, 16);
         ooo_mgr_zuc_reset(state->zuc_nea6_ooo, 16);
         ooo_mgr_zuc_reset(state->zuc_nia6_ooo, 16);
-        ooo_mgr_zuc_reset(state->zuc_nca6_enc_ooo, 4);
-        ooo_mgr_zuc_reset(state->zuc_nca6_dec_ooo, 4);
+        ooo_mgr_zuc_reset(state->zuc_nca6_enc_ooo, 16);
+        ooo_mgr_zuc_reset(state->zuc_nca6_dec_ooo, 16);
 
         /* Init HMAC/SHA1 out-of-order fields */
         ooo_mgr_hmac_sha1_reset(state->hmac_sha_1_ooo, AVX512_NUM_SHA1_LANES);